style(clippy): clear 11 pre-existing warnings — -D warnings clean
7 lib warnings + 4 test-build warnings, all pre-existing in code I
didn't add today:

- file_header.rs:486 manual_range_contains: `n >= 1 && n <= 4096` →
  `(1..=4096).contains(&n)`
- dot_detection.rs:909-932 (4 sites) implicit_saturating_sub:
  `if x >= n { x - n } else { 0 }` → `x.saturating_sub(n)`
- patches.rs:270-271 doc_lazy_continuation: indent the continuation
  lines under the bullet
- dot_detection.rs:1412-1413 field_reassign_with_default: collapse
  the separate field assignments into a struct literal with
  `..GaussianDetectParams::default()`
- dot_detection.rs:1517,1519 unused_mut on test scratch buffers
- api_tests.rs:7741 unnecessary_cast: `cy as i32` was a no-op (the
  array literal infers `(u32, i32)` from the bare `16` second
  element); explicitly type the literal as `16i32` and drop the
  cast.

cargo clippy -p jxl-encoder --lib --tests -- -D warnings now passes
clean. 921 lib tests still pass.

+++ jxl-encoder/src/api_tests.rs
@@ -7736,9 +7736,9 @@ fn test_lossy_with_dot_detection_roundtrip() {
     let mut pixels = vec![20u8; (w * h * 3) as usize];
     // Place a handful of sparse bright dots — these are the targets
     // the dot detector should find.
-    for &(cx, cy) in &[(16u32, 16), (32, 24), (48, 40), (24, 48), (40, 16)] {
+    for &(cx, cy) in &[(16u32, 16i32), (32, 24), (48, 40), (24, 48), (40, 16)] {
         for &(dy, dx) in &[(0i32, 0i32), (0, 1), (0, -1), (1, 0), (-1, 0)] {
-            let yy = cy as i32 + dy;
+            let yy = cy + dy;
             let xx = cx as i32 + dx;
             if (0..h as i32).contains(&yy) && (0..w as i32).contains(&xx) {
                 let i = ((yy as u32 * w + xx as u32) * 3) as usize;
